+++ js/dialogs.js
@@ -1,5 +1,18 @@
 import { isEscapeKey } from './util.js';
 
+const ALERT_SHOW_TIME = 5000;
+const alertTemplate = document.querySelector('#data-error').content;
+
+export const showAlert = () => {
+  const alertContainer = alertTemplate.cloneNode(true);
+  const errorMessage = alertContainer.querySelector('.data-error');
+  document.body.append(errorMessage);
+
+  setTimeout(() => {
+    errorMessage.remove();
+  }, ALERT_SHOW_TIME);
+};
+
 const removeMessage = (message, onAction) => {
   message.remove();
   document.removeEventListener('keydown', onAction);
@@ -23,3 +36,4 @@ export const showMessage = (template, messageClass) => {
     removeMessage(message, onDocumentKeyDown);
   });
 };
+
